Create util classes. Move functions around.
Remove recoil notification from jewelry count plugin.
Create WildcardMatchLoader in utils and remove all local instances of it from other plugins.
