+++ runelite-client/src/main/java/net/runelite/client/plugins/jewellerycount/JewelleryCountPlugin.java
@@ -25,13 +25,9 @@
 package net.runelite.client.plugins.jewellerycount;
 
 import javax.inject.Inject;
-import net.runelite.api.ChatMessageType;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.ui.overlay.Overlay;
-import com.google.common.eventbus.Subscribe;
 import com.google.inject.Provides;
-import javax.inject.Inject;
-import net.runelite.api.events.ChatMessage;
 import net.runelite.client.Notifier;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.plugins.PluginDescriptor;
@@ -60,17 +56,5 @@ public class JewelleryCountPlugin extends Plugin
 	JewelleryCountConfig getConfig(ConfigManager configManager)
 	{
 		return configManager.getConfig(JewelleryCountConfig.class);
-	}
-
-	@Subscribe
-	public void onChatMessage(ChatMessage event)
-	{
-		if (event.getType() == ChatMessageType.SERVER)
-		{
-			if (config.recoilNotification() && event.getMessage().contains("<col=7f007f>Your Ring of Recoil has shattered.</col>"))
-			{
-				notifier.notify("Your Ring of Recoil has shattered");
-			}
-		}
 	}
 }
