+++ runelite-client/src/main/java/net/runelite/client/util/WildcardMatchLoader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Adam <Adam@sigterm.info>
+ * Copyright (c) 2018, Tomas Slusny <slusnucky@gmail.com>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,24 +22,40 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.grounditems;
+package net.runelite.client.util;
 
-import java.util.Arrays;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import org.junit.Test;
+import com.google.common.base.Strings;
+import com.google.common.cache.CacheLoader;
+import java.util.List;
+import javax.annotation.Nonnull;
 
-public class WildcardMatchLoaderTest
+public class WildcardMatchLoader extends CacheLoader<String, Boolean>
 {
-	@Test
-	public void testLoad()
+	private final List<String> nameFilters;
+
+	public WildcardMatchLoader(List<String> nameFilters)
 	{
-		WildcardMatchLoader loader = new WildcardMatchLoader(Arrays.asList("rune*", "Abyssal whip"));
-		assertTrue(loader.load("rune pouch"));
-		assertTrue(loader.load("Rune pouch"));
-		assertFalse(loader.load("Adamant dagger"));
-		assertTrue(loader.load("Runeite Ore"));
-		assertTrue(loader.load("Abyssal whip"));
-		assertFalse(loader.load("Abyssal dagger"));
+		this.nameFilters = nameFilters;
+	}
+
+	@Override
+	public Boolean load(@Nonnull final String key)
+	{
+		if (Strings.isNullOrEmpty(key))
+		{
+			return false;
+		}
+
+		final String filteredName = key.trim();
+
+		for (final String filter : nameFilters)
+		{
+			if (WildcardMatcher.matches(filter, filteredName))
+			{
+				return true;
+			}
+		}
+
+		return false;
 	}
 }
